+++ src/ui/switchboard/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""
+Switchboard UI Package
+Contains Switchboard-related UI components
+"""
+
+from .switchboard_widget import SwitchboardWidget
+
+__all__ = ['SwitchboardWidget']